TL review of day 2 homework

+++ MPAGSCipher/CommandLine.hpp
@@ -13,6 +13,6 @@ bool processCommandLine(
  std::string& inputFileName,
  std::string& outputFileName,
  bool& encrypt,
- int& key);
+ size_t& key);
 
 #endif
